refactor(scripts): py2/3-safe syntax, ui root (1/2)

+++ Lead-Client/pack/root/uicommon.py
@@ -412,7 +412,7 @@ class MoneyInputDialog(ui.ScriptWindow):
 		money = 0
 		if text and text.isdigit():
 			try:
-				money = long(text)
+				money = int(text)
 			except ValueError:
 				money = 199999999
